wip: strip handler out of event.rs
